+++ src/assets/ba_data/python/baenv.py
@@ -58,7 +58,7 @@ logger = logging.getLogger('baenv')
 
 # Build number and version of the ballistica binary we expect to be
 # using.
-TARGET_BALLISTICA_BUILD = 22422
+TARGET_BALLISTICA_BUILD = 22426
 TARGET_BALLISTICA_VERSION = '1.7.44'