+++ src/ballistica/base/input/device/keyboard_input.h
@@ -18,7 +18,7 @@ class KeyboardInput : public InputDevice {
   ~KeyboardInput() override;
   auto HandleKey(const SDL_Keysym* keysym, bool down) -> bool;
   void ApplyAppConfig() override;
-  auto GetRawDeviceName() -> std::string override;
+  auto DoGetDeviceName() -> std::string override;
   void ResetHeldStates() override;
   auto left_key_assigned() const { return left_key_assigned_; }
   auto right_key_assigned() const { return right_key_assigned_; }