+++ src/ballistica/base/input/device/touch_input.h
@@ -29,7 +29,7 @@ class TouchInput : public InputDevice {
   enum class ActionControlType { kButtons, kSwipe };
 
  protected:
-  auto GetRawDeviceName() -> std::string override;
+  auto DoGetDeviceName() -> std::string override;
 
  private:
   void UpdateDPad();