+++ src/ballistica/core/logging/logging.h
@@ -22,19 +22,22 @@ class Logging {
   void Log(LogName name, LogLevel level, const char* msg);
   void Log(LogName name, LogLevel level, char* msg);
 
-  void ApplyBaEnvConfig();
-
   /// Log call variant taking a call returning a string instead of a string
-  /// directly. This can be useful for log strings requiring significant
-  /// effort to construct, as the call will be skipped if the log level is
-  /// not currently visible.
+  /// directly. This is recommended for log strings requiring significant
+  /// effort to construct, as said construction will be skipped if the log
+  /// level is not currently enabled.
   template <typename C>
   void Log(LogName name, LogLevel level, C getmsgcall) {
-    if (!LogLevelEnabled(name, level)) {
-      return;
+    // Make sure provided lambdas return std::string; otherwise it would be
+    // an easy mistake to return a char* to invalid function-local memory.
+    static_assert(std::is_same<std::string, decltype(getmsgcall())>::value,
+                  "Lambda must return std::string");
+    if (LogLevelEnabled(name, level)) {
+      Log_(name, level, getmsgcall());
     }
-    Log(name, level, getmsgcall());
   }
+
+  void ApplyBaEnvConfig();
 
   /// Grab current Python logging levels for all logs we use internally. If
   /// any changes are made at runtime to Python logging levels that we use,
@@ -54,20 +57,6 @@ class Logging {
     return static_cast<int>(log_levels_[static_cast<int>(name)]);
   }
 
-  /// Write a message to the log. Intended for logging use in C++ code. This
-  /// is safe to call by any thread at any time as long as core has been
-  /// inited. In general it simply passes through to the equivalent Python
-  /// logging call: logging.info, logging.warning, etc.
-  ///
-  /// Be aware that Log() calls made before babase is imported will be
-  /// stored and submitted all at once to Python once babase is imported
-  /// (with a [HELD] prefix). Ballistica's log/print redirection gets
-  /// finalized at that point and this system ensures all C++ Log() calls
-  /// ever made will be routed through the app, visible in in-app consoles,
-  /// etc. Note that direct Python logging calls or prints occurring before
-  /// babase is imported may not be visible in the app for that same reason.
-  void Log_(LogName name, LogLevel level, const std::string& msg);
-
   /// Send a log message to the in-app console, platform-specific logs, etc.
   /// This generally should not be called directly but instead wired up to
   /// log messages coming through the Python logging system.
@@ -85,6 +74,20 @@ class Logging {
   auto v1_cloud_log_full() const { return v1_cloud_log_full_; }
 
  private:
+  /// Write a message to the log. Intended for logging use in C++ code. This
+  /// is safe to call by any thread at any time as long as core has been
+  /// inited. In general it simply passes through to the equivalent Python
+  /// logging call: logging.info, logging.warning, etc.
+  ///
+  /// Be aware that Log() calls made before babase is imported will be
+  /// stored and submitted all at once to Python once babase is imported
+  /// (with a [HELD] prefix). Ballistica's log/print redirection gets
+  /// finalized at that point and this system ensures all C++ Log() calls
+  /// ever made will be routed through the app, visible in in-app consoles,
+  /// etc. Note that direct Python logging calls or prints occurring before
+  /// babase is imported may not be visible in the app for that same reason.
+  void Log_(LogName name, LogLevel level, const std::string& msg);
+
   LogLevel log_levels_[static_cast<int>(LogName::kLast)]{};
   bool did_put_v1_cloud_log_{};
   bool v1_cloud_log_full_{};